update based on test comments
