+++ packages/contact-center/cc-components/src/components/task/CallControlCAD/call-control-cad.styles.scss
@@ -27,7 +27,7 @@
   flex-flow: row wrap;
   justify-content: space-between;
   width: 100%;
-  max-height: 180px;
+  max-height: 11.25rem;
   overflow-y: auto;
   margin-top: 0.75rem;
 
@@ -37,24 +37,6 @@
     box-sizing: border-box;
     padding: 0.125rem 0;
     line-height: 1.75rem;
-
-    .global-variable-label {
-      display: block;
-      font-size: 0.875rem;
-      font-weight: bold;
-      color: var(--mds-color-theme-text-secondary-normal);
-      overflow: hidden;
-      white-space: nowrap;
-      text-overflow: ellipsis;
-    }
-
-    .global-variable-value {
-      display: block;
-      font-size: 0.875rem;
-      font-weight: 250;
-      color: var(--mds-color-theme-text-primary-normal);
-      word-break: break-word;
-    }
   }
 }
